+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.x branch
 ## 1.4 branch
 ### 1.4.3
+* added `array_unique_recursive()` global function;
 * added `Exceptionist::methodExists()` method;
 * fixed compatibility with PHP 7.4 and phpunit 9.